Uses development version of Jesse-Islam/riskRegression
It seems to be working with my version of riskRegression, with the function acting as a part of the package.

The issue with 1- absolute risk I was worried about is not longer an issue. Turns out It was wrong before anyway. predictRisk attempts to output the probability of the event occuring